Unified Latte integers as signed 32-bit values.
  - Lexer intentionally parses numeric tokens as i64 for overflow diagnosis.
  - Parser performs the single checked i64 → i32 conversion with the literal’s source span.
  - AST, typed AST, and custom IR now store integers as i32.
  - LLVM emission uses signed i32 constants directly.
  - Removed the obsolete frontend/backend width-mismatch TODO.
  - Fixed build.rs so grammar changes regenerate the parser.

  Added regressions:

  - 2147483647 compiles and runs.
  - 2147483648 is rejected with: “Integer literal is outside the 32-bit signed range.”

+++ src/ir.rs
@@ -73,7 +73,7 @@ impl Phi {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Value {
-    Int(i64),
+    Int(i32),
     String(String),
     Bool(bool),
     Call(VariableId, Vec<ValueId>),
